Add menuStatus model to controls to track open state
This allows all child views of each control view (font-family, font-variant, and
font-size) to always know and react to the current state of the menu (open or
closed). This allows us to hide the DefaultFontButton if the menu is open. It
also allows triggering emitter events for open-menu and close-open-menus instead
of the non-declarative toggle-dropdown event.