General cleanup
- Adds nvmrc (going from node6 to node12)
- Misc fixes for npm packages / js tests
- Misc fixes for php, some of the tests weren't passing with production
code so there are some fixes that may need real testing.

Specifically concerned about the whitelist caching. It's dependent on the
whitelist being static and not dynamic like we plan to do with the non latin
fonts about to be added. Probably need to take a better run at this
tomorrow.