Google Provider: Fix allowed font test
The body text property of a font is set to `false` when it's not allowed
as a body font, but is allowed as a header font. The test font of Antic
wasn't allowed as either so shouldn't have been allowed but because it
wasn't explicitly a header font, the logic returned true.

This change updates the test font to one that meets the criteria of
being a header font but not a body font.

It's possible that we should update this logic to so that both
conditions should be true. The body text property will be set to true
for all fonts that aren't heading fonts currently.